==> REST[ES]_donors/Donor8/2_Normalize_for_donor.Rmd <==
---
title: "Normalize Data"
author: "Wesley Burr"
date: "29/03/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
```

## Normalize Data

We will load each of the data files, identify the 'primary' peak for
Bromobenze, then normalize all Areas to be relative areas to the
reference standard.

```{r}
load("/home/wburr/Rushali_REST/Donor8/D8.rda")
D8m$Area<-as.numeric(D8m$Area)
bromo <- D8m %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(Area == max(Area))
split_Donor <- D8m %>% 
                group_by(Sample) %>%
                group_split()
for(j in 1:length(split_Donor)) {
  split_Donor[[j]]$Area <- split_Donor[[j]]$Area / unlist(bromo[j, "Area"])
}
D8m <- do.call("rbind", split_Donor)
save(file = "/home/wburr/Rushali_REST/Donor8/D8n.rda", D8m)
```
